Add files via upload
Programming Application: IT Department Program
The following project was written in Java, following the best practices and all the possible features, it describes an online retail shop that needs a program to help manage their assets. 
It was divided into classes and methods to meet all the requirements, which each one will be explained individually.